requirements.txt
torch>=2.0.0
numpy>=1.21.0
asyncio-dgram>=2.1.0
aiohttp>=3.8.0
cryptography>=3.4.0
msgpack>=1.0.0
psutil>=5.8.0
networkx>=2.6.0
websockets>=10.0
uvloop>=0.17.0; sys_platform != "win32"
pyyaml>=6.0